Build fix for Win64 after r16521.
TBR=verwaest@chromium.org

Review URL: https://codereview.chromium.org/23850003

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@16524 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/platform/socket.cc
@@ -128,7 +128,8 @@ bool Socket::Connect(const char* host, const char* port) {
   for (struct addrinfo* ai = info; ai != NULL; ai = ai->ai_next) {
     // Try to connect using this addr info.
     while (true) {
-      result = ::connect(native_handle_, ai->ai_addr, ai->ai_addrlen);
+      result = ::connect(
+          native_handle_, ai->ai_addr, static_cast<int>(ai->ai_addrlen));
       if (result == 0) {
         freeaddrinfo(info);
         return true;
